added winston for logging

+++ src/controllers/userController.ts
@@ -4,9 +4,11 @@ import {User} from "../models/User";
 import {generateFakeUserData} from "../utils/generateFakeUserData";
 import {OmniService} from "../services/external/omni/OmniService";
 import {NipClient} from "../services/external/nip/NipClient";
+import logger from '../utils/logger';
 
 
 const userService = new UserService();
+const nipClient: NipClient = new NipClient();
 
 export async function getUserHandler(
     req: Request,
@@ -16,6 +18,7 @@ export async function getUserHandler(
     try {
         let user: User | null = await userService.getAvailableUser();
         if (!user) {
+            logger.warn('No available user found');
             res.status(404).json({message: "No users available"});
             return;
         }
@@ -66,18 +69,20 @@ export async function unlockUserHandler(
         const {email} = req.body;
 
         if (!email || typeof email !== "string") {
+            logger.warn(`Couldn't find a user with the email ${email}`);
             res.status(400).json({message: "Email query parameter of string is required"});
             return
         }
-        let user: User | null = await userService.getUserByEmail(email);
 
+        let user: User | null = await userService.getUserByEmail(email);
         if (!user?.reserved) {
+            logger.warn(`User with the email ${email} is not reserved`);
             res.status(400).json({message: "User is not reserved"});
             return
         }
 
         await userService.unlockUser(user.id);
-
+        logger.info(`Successfully unlocked user with email ${email}`);
         res.status(204).send();
     } catch (error) {
         next(error);
@@ -98,8 +103,12 @@ export async function removeUserHandler(
         }
 
         let user: User | null = await userService.getUserByEmail(email);
+        if (!user?.reserved) {
+            logger.warn(`User with email ${email} is not reserved`);
+            res.status(400).json({ message: "User is not reserved" });
+            return;
+        }
 
-        let nipClient: NipClient = new NipClient();
         const {memberId} = await nipClient.getCustomer(email);
 
         const deleteMemberResponse = await nipClient.deleteMember(memberId);
